+++ webots_ros2_driver/src/plugins/static/Ros2RangeFinder.cpp
@@ -25,13 +25,18 @@ namespace webots_ros2_driver {
     mIsEnabled = false;
     mRangeFinder = wb_robot_get_device(parameters["name"].c_str());
 
+    mCameraInfoSuffix = parameters.count("cameraInfoSuffix") ? parameters["cameraInfoSuffix"] : "/camera_info";
+    mImageSuffix = parameters.count("imageSuffix") ? parameters["imageSuffix"] : "/image";
+    mPointCloudSuffix = parameters.count("pointCloudSuffix") ? parameters["pointCloudSuffix"] : "/point_cloud";
+
     assert(mRangeFinder != 0);
 
     const int width = wb_range_finder_get_width(mRangeFinder);
     const int height = wb_range_finder_get_height(mRangeFinder);
 
     // Image publisher
-    mImagePublisher = mNode->create_publisher<sensor_msgs::msg::Image>(mTopicName, rclcpp::SensorDataQoS().reliable());
+    mImagePublisher =
+      mNode->create_publisher<sensor_msgs::msg::Image>(mTopicName + mImageSuffix, rclcpp::SensorDataQoS().reliable());
     mImageMessage.header.frame_id = mFrameName;
     mImageMessage.height = height;
     mImageMessage.width = width;
@@ -42,7 +47,7 @@ namespace webots_ros2_driver {
 
     // CameraInfo publisher
     mCameraInfoPublisher =
-      mNode->create_publisher<sensor_msgs::msg::CameraInfo>(mTopicName + "/camera_info", rclcpp::SensorDataQoS().reliable());
+      mNode->create_publisher<sensor_msgs::msg::CameraInfo>(mTopicName + mCameraInfoSuffix, rclcpp::SensorDataQoS().reliable());
     mCameraInfoMessage.header.stamp = mNode->get_clock()->now();
     mCameraInfoMessage.header.frame_id = mFrameName;
     mCameraInfoMessage.height = height;
@@ -59,8 +64,8 @@ namespace webots_ros2_driver {
                             1.0,          0.0};
 
     // Point cloud publisher
-    mPointCloudPublisher =
-      mNode->create_publisher<sensor_msgs::msg::PointCloud2>(mTopicName + "/point_cloud", rclcpp::SensorDataQoS().reliable());
+    mPointCloudPublisher = mNode->create_publisher<sensor_msgs::msg::PointCloud2>(mTopicName + mPointCloudSuffix,
+                                                                                  rclcpp::SensorDataQoS().reliable());
     mPointCloudMessage.header.frame_id = mFrameName;
     mPointCloudMessage.fields.resize(3);
     mPointCloudMessage.fields[0].name = "x";
@@ -118,6 +123,7 @@ namespace webots_ros2_driver {
     auto image = wb_range_finder_get_range_image(mRangeFinder);
     if (image) {
       mImageMessage.header.stamp = mNode->get_clock()->now();
+      mCameraInfoMessage.header.stamp = mImageMessage.header.stamp;
       memcpy(mImageMessage.data.data(), image, mImageMessage.data.size());
       mImagePublisher->publish(mImageMessage);
     }