+++ webots_ros2_mavic/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package webots_ros2_mavic
 2023.1.0 (2023-XX-XX)
 ------------------
 * Clean simulation reset in launch file.
+* Update driver node to new WebotsController node.
 
 2023.0.2 (2023-02-07)
 ------------------